[WEJBHTTP-139] Refactoring - making all ByteOutputs wrapping OutputStreams unflushable
