--- common/src/main/java/org/wildfly/httpclient/common/NoFlushByteOutput.java
@@ -1,64 +1,0 @@
-/*
- * JBoss, Home of Professional Open Source.
- * Copyright 2017 Red Hat, Inc., and individual contributors
- * as indicated by the @author tags.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.wildfly.httpclient.common;
-
-import java.io.IOException;
-
-import org.jboss.marshalling.ByteOutput;
-
-/**
- * An output stream that ignores flushes. The marshsaller will flush when it is done, which
- * results in two frames on the wire. By ignoring the flush only one frame is sent for
- * each message.
- *
- * @author Stuart Douglas
- */
-public class NoFlushByteOutput implements ByteOutput {
-
-    private final ByteOutput delegate;
-
-    public NoFlushByteOutput(ByteOutput delegate) {
-        this.delegate = delegate;
-    }
-
-    @Override
-    public void write(int b) throws IOException {
-        delegate.write(b);
-    }
-
-    @Override
-    public void write(byte[] b) throws IOException {
-        delegate.write(b);
-    }
-
-    @Override
-    public void write(byte[] b, int off, int len) throws IOException {
-        delegate.write(b, off, len);
-    }
-
-    @Override
-    public void close() throws IOException {
-        delegate.close();
-    }
-
-    @Override
-    public void flush() throws IOException {
-        //ignore
-    }
-}
